Native support for bitfields, first stage
In CompCert C and Clight, this commit adds a "bitfield" designator to l-value expressions and to the results of their evaluation.

Bitfield designators are either
- `Full`: no bit field, the value accessed is that contained in memory at the given block and offset
- `Bits carrier pos width`: the value accessed is `width` bits starting at bit position `pos` in the integer of size `carrier` contained in memory at the given block and offset.

During Csharpminor generation, bitfield accesses are implemented by integer shifts and masks.

Currently, all l-value expressions produce a `Full` bitfield, so the `Bits` case is never exercised.  This will change next when the layout algorithm for struct members is extended to deal with bitfield members.