Revise semantics of assignment w.r.t. the return value of an assignment
The value of `lhs = rhs` used to be the value of `rhs` cast to the
type of `lhs`.  This is incorrect if `lhs` denotes a bitfield in a
struct or union.  E.g. if `int f: 2`, `x.f = 7` evaluates to `-1`, not `7`.

In this commit we modify `assign_loc` and `store_bitfield` to return
the value of the assignment expression.  We use this in the semantics
of assignment expressions in Csem.

We then adapt SimplExpr to the revised semantics of assignment return
values.  The generated Clight code now correctly zero- or sign-extends
the assigned value if the lhs is a bitfield.