Ignore the volatile modifier on bitfield accesses
Change the Compcert C semantics to ignore the volatile modifier on
access to bit fields, just like it ignores it on accesses by value to
structs and unions, and access them as if they were not volatile.
(Before, volatile bit field accesses were undefined behavior in Csem.)

Change the SimplExpr pass to not generate EF_vload and EF_vstore
builtins for accesses to bit fields of volatile type, just like it
does not generate them for accesses to volatile structs and unions.